requirements.txt
swarms>=0.3.0
pylint>=2.17.0
mypy>=1.5.1
bandit>=1.7.5
flake8>=6.1.0
radon>=6.0.1
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.2
python-dotenv>=1.0.0
asyncio>=3.4.3
aiohttp>=3.9.1
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
httpx>=0.25.2
pytest-mock>=3.12.0
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
alembic>=1.12.1
asyncpg>=0.29.0 